+++ packages/parser/src/envelopes/event-bridge.ts
@@ -17,7 +17,7 @@ export const EventBridgeEnvelope = {
     return Envelope.parse(EventBridgeSchema.parse(data).detail, schema);
   },
 
-  safeParse<T extends ZodSchema>(data: unknown, schema: T): ParsedResult {
+  safeParse<T extends ZodSchema>(data: unknown, schema: T): ParsedResult<unknown, z.infer<T>> {
     const parsedEnvelope = EventBridgeSchema.safeParse(data);
 
     if (!parsedEnvelope.success) {