+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0](https://github.com/aws-powertools/powertools-lambda-typescript/compare/v2.10.0...v2.11.0) (2024-11-20)
+
+
+### Bug Fixes
+
+* **parser:** add aws region to kinesis event ([#3260](https://github.com/aws-powertools/powertools-lambda-typescript/issues/3260)) ([246f132](https://github.com/aws-powertools/powertools-lambda-typescript/commit/246f13253bdba1f6963cf53605b0ae10698f063e))
+* **parser:** event type literal for selfManagedKafka ([#3325](https://github.com/aws-powertools/powertools-lambda-typescript/issues/3325)) ([5350afe](https://github.com/aws-powertools/powertools-lambda-typescript/commit/5350afed92d02b7b8d47f387705f70c59deeeb65))
+* **parser:** fix cause errors nested structure ([#3250](https://github.com/aws-powertools/powertools-lambda-typescript/issues/3250)) ([1ff97cb](https://github.com/aws-powertools/powertools-lambda-typescript/commit/1ff97cb758b2e44a76ce108d8e54c9335088abba))
+
+
+
+
+
 # [2.9.0](https://github.com/aws-powertools/powertools-lambda-typescript/compare/v2.8.0...v2.9.0) (2024-10-07)
 
 **Note:** Version bump only for package @aws-lambda-powertools/parser